Tests/LibWeb/Text/expected/textarea-arrow-navigation-with-letter-spacing.txt

Down from offset 3: position=8
Up from offset 9: position=5
